Crate hyperion_syscall

Modules

err

fs

id

net

Functions

accept

accept a connection on a socket

bind

bind a name to a socket

close

close a file

connect

connect to a socket

dup

duplicate a file descriptor

exit

exit the process with a code

futex_wait

futex wait if value at addr is val

futex_wake

wake num threads that are sleeping on this addr

get_pid

get the current process id

get_tid

get the current thread id

listen

start listening for connections on a socket

log

print a string into kernel logs

map_file

map file contents to memory (mmap)

metadata

file metadata (stat)

nanosleep

context switch from this process and switch back when nanos nanoseconds have passed

nanosleep_until

context switch from this process and switch back when timestamp() > deadline_nanos

open

open a file

open_dir

open a directory

palloc

allocate physical pages and map to heap

pfree

deallocate physical pages and unmap from heap

pthread_spawn

spawn a new pthread for the same process

read

read from a file

recv

read data from a socket

rename

rename the current process

seek

file position seek (fseek)

send

send data to a socket

socket

create a socket

syscall_0⚠

Safety

syscall_1⚠

Safety

syscall_2⚠

Safety

syscall_3⚠

Safety

syscall_4⚠

Safety

syscall_5⚠

Safety

timestamp

u128 nanoseconds since boot

unmap_file

unmap device/file mapped memory (munmap)

write

write into a file

yield_now

context switch from this process, no guarantees about actually switching