Crate hyperion_syscall

Modules

err

Functions

exit

exit the process with a code

log

print a string into kernel logs

nanosleep

context switch from this process and switch back when nanos nanoseconds have passed

nanosleep_until

context switch from this process and switch back when timestamp() > deadline_nanos

palloc

allocate physical pages and map to heap

pfree

deallocate physical pages and unmap from heap

pthread_spawn

spawn a new pthread for the same process

recv

read data from a PID based single naïve IPC channel

rename

rename the current process

send

send data to a PID based single naïve IPC channel

syscall_0⚠

Safety

syscall_1⚠

Safety

syscall_2⚠

Safety

syscall_3⚠

Safety

syscall_4⚠

Safety

syscall_5⚠

Safety

timestamp

u128 nanoseconds since boot

yield_now

context switch from this process, no guarantees about actually switching